fix: conditionally hide "Downloads" title
- title is hidden for bouquet/conservation projects with no tax receipt
- minor refactors
a. extracts `Certificates` to a separate file
b. renames `disableCertificate` to `shouldDisableCertificate` (clarity)
c. exports `shouldDisableCertificate`for reuse within `AccountRecord`
d. Reorders components within `AccountRecord.tsx`
